move products app to the right package
