+++ reactive-architecture/solutions/products/src/main/java/com/redhat/training/PricesService.java
@@ -1,4 +1,4 @@
-package org.acme.quickstart;
+package com.redhat.training;
 
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 // import org.jboss.resteasy.annotations.jaxrs.QueryParam;
