+++ reactive-architecture/solutions/products/src/main/java/com/redhat/training/ProductPriceHistory.java
@@ -1,4 +1,4 @@
-package org.acme.quickstart;
+package com.redhat.training;
 
 import java.util.List;
