--- reactive-architecture/solutions/products/src/test/java/org/acme/quickstart/NativeGreetingResourceIT.java
@@ -1,9 +1,0 @@
-package org.acme.quickstart;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeGreetingResourceIT extends GreetingResourceTest {
-
-    // Execute the same tests but in native mode.
-}
